Fix layout issues for main image and thumbnails.
This adds a bit of CSS to fix the main image layout. First, the root
div needs height 100%. In the Build UI demo we were using Next whose
root element was `#__next`, but for Vite it's `#root`.

Next, the images in this project are 853 x 480 pixels, so on fullscreen
when the container was 1280 they wouldn't take up the space. I added
a width of 100% and also `flex-shrink: 0` to force them to take up
the full space available.

Finally, I added `flex-shrink-0` to the buttons in the thumbnails
so Safari sizing would work properly.